project/requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
dominate==2.6.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Testing==0.8.1
Flask-WTF==1.0.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
